requirements.txt
asyncgui==0.6.3
asynckivy==0.6.4
bcrypt==5.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
docutils==0.22.3
filetype==1.2.0
idna==3.11
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivy_deps.gstreamer==0.3.4
kivy_deps.sdl2==0.8.0
kivymd @ file:///C:/Users/igorm/OneDrive/%C3%81rea%20de%20Trabalho/Cognitive/KivyMD
materialshapes==0.3
materialyoucolor==2.0.10
pillow==12.0.0
psycopg2==2.9.11
pycairo==1.28.0
Pygments==2.19.2
pypiwin32==223
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
requests==2.32.5
resend==2.19.0
typing_extensions==4.15.0
urllib3==2.5.0
